+++ examples/node/4-store-read-events/index.ts
@@ -1,9 +1,15 @@
-import { DdcClient, Link, DagNode, TESTNET, DagNodeUri } from '@cere-ddc-sdk/ddc-client';
+import { DdcClient, Link, DagNode, DagNodeUri } from '@cere-ddc-sdk/ddc-client';
+import type { ClusterId } from '@cere-ddc-sdk/blockchain';
 
 /**
  * The wallet should have enough CERE to pay for the transaction fees
  */
 const user = 'hybrid label reunion only dawn maze asset draft cousin height flock nation';
+
+/**
+ * The DDC cluster where the bucket lives
+ */
+const clusterId: ClusterId = '0x825c4b2352850de9986d9d28568db6f0c023a1e3';
 
 /**
  * The DDC bucket where the file will be stored
@@ -18,7 +24,13 @@ const streamCnsName = 'events-stream-example';
 /**
  * Create a DDC client instance and connect it to DDC TESTNET
  */
-const client = await DdcClient.create(user, { ...TESTNET, logLevel: 'silent' });
+const client = await DdcClient.create(user, {
+  blockchain: 'testnet',
+  clusterId,
+  storageUrl: 'https://storage.testnet.dragon-1.xyz',
+  cdnUrl: 'https://cdn.testnet.dragon-1.xyz',
+  logLevel: 'silent',
+});
 
 /**
  * Get the previous event root attached to the stream CNS record